% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probs_comb.R
\name{calculate_prob_of_comb}
\alias{calculate_prob_of_comb}
\title{Calculate probabilities of modifiers and modifier combinations occurring}
\usage{
calculate_prob_of_comb(data, modifiers)
}
\arguments{
\item{data}{data frame as extracted from Filemaker (potentially cleaned)}

\item{modifiers}{name of the modifiers that should be included}
}
\value{
Function returns a dataframe with observed probabilities for each combination in the dataset
}
\description{
Calculate probabilities of modifiers and modifier combinations occurring
}
