% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_extract.R
\name{clean_function}
\alias{clean_function}
\title{Clean datasets after they come out of Filemaker}
\usage{
clean_function(
  data,
  clean.value.levels = TRUE,
  clean.col.names = TRUE,
  exclude.unknown.inds = FALSE,
  exclude.unknown.gestures = FALSE,
  exclude.incomplete.coding = FALSE,
  exclude.unknown.direction = FALSE,
  exclude.goal.unmet = FALSE,
  exclude.no.response.waiting = FALSE,
  exclude.no.persistence = FALSE,
  exclude.unknown.goal = FALSE
)
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{clean.value.levels}{TRUE or FALSE; should value names be cleaned to be more easily comprehensible for all columns?}

\item{clean.col.names}{TRUE or FALSE; should the column names be cleaned to more easily understandable}

\item{exclude.unknown.inds}{TRUE or FALSE; should unknown Signaller be excluded from the data?}

\item{exclude.unknown.gestures}{TRUE or FALSE; should cases be excluded if the Gesture_record is ["None","Unk","Time_ex", "no_value", "Unclear", "Pot_new"], or if the Directed_to is ["No", "Sev_pot","Unk"]}

\item{exclude.incomplete.coding}{TRUE or FALSE; should incomplete coding be excluded from the data?}

\item{exclude.unknown.direction}{TRUE or FALSE; should unknown direction be excluded from the data?}

\item{exclude.goal.unmet}{TRUE or FALSE; should unmet goals be excluded from the data?}

\item{exclude.no.response.waiting}{TRUE or FALSE; should lacking response waiting be excluded from the data?}

\item{exclude.no.persistence}{TRUE or FALSE; should lacking persistence be excluded from the data?}

\item{exclude.unknown.goal}{TRUE or FALSE; should unknown goal be excluded from the data?}
}
\value{
Function returns a cleaned version of the dataframe that is entered
}
\description{
Clean datasets after they come out of Filemaker
}
