% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_to_levels.R
\name{colnames_to_levels}
\alias{colnames_to_levels}
\title{Helper functions, adds column name to every level of the dataframe}
\usage{
colnames_to_levels(modifier.data)
}
\arguments{
\item{modifier.data}{data frame with columns representing modifiers}
}
\value{
Same data frame, but column names added to levels
}
\description{
Helper functions, adds column name to every level of the dataframe
}
