% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_modifiers.R
\name{conditional_modifiers}
\alias{conditional_modifiers}
\title{Calculate conditional probabilities of modifier levels across modifiers within the same gesture action, to see if some can be removed}
\usage{
conditional_modifiers(data, modifiers, threshold = 0.98)
}
\arguments{
\item{data}{data frame as extracted from Filemaker (potentially cleaned)}

\item{modifiers}{name of the modifiers that should be included}

\item{threshold}{above which conditional probability (both ways) should modifier levels be considered to be strongly connected?}
}
\value{
Function returns a dataframe with those modifier level combinations that have above-threshold conditional probabilities both ways
}
\description{
For example, if in a Reach gesture, Hand Flexion and Finger Flexion occur together 100% of the time, we can just treat them as the same
}
