% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_from_clip.R
\name{date_from_clip}
\alias{date_from_clip}
\title{Helper functions to extract info from Clip and Communication Numbers}
\usage{
date_from_clip(clip_column)
}
\arguments{
\item{clip_column}{vector with clip names}
}
\value{
returns Date, ClipNr, Coder, Group, and Species as vectors
}
\description{
Helper functions to extract info from Clip and Communication Numbers
}
