% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_detection.R
\name{error_detection}
\alias{error_detection}
\title{Function takes the gesture definition file as produced by Kirsty with everybody's input, and detects cases where those definitions for possible modifiers are violated}
\usage{
error_detection(data, col_names = "renamed")
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{col_names}{defines whether the 'original' names were used (i.e., as the come out of Filemaker) or it was 'renamed' using other functions I wrote}
}
\value{
Function returns data frame with potential errors and their description, including coder, communication number, and a description of what should be checked
}
\description{
Potential errors here only concern modifiers that are not defined in the protocols but occur. These should be reported back to the coder, to be changed in ELAN and FileMaker
}
