% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesture_context_analytical.R
\name{gesture_context_analytical}
\alias{gesture_context_analytical}
\title{Compares the distribution of Goals of the same gesture action against what would be expected from unconditional probabilities}
\usage{
gesture_context_analytical(
  data,
  threshold = 10,
  goal_met = TRUE,
  reduce.goals = TRUE,
  reduce.gestures = TRUE
)
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{threshold}{minimum number of times a gesture has to be marked as a certain goal before it is included}

\item{goal_met}{Do Goal_met have to be 'yes'?}

\item{reduce.goals}{if TRUE, some goals will be combined (e.g., GroomMe and GroomYou become 'Groom')}

\item{reduce.gestures}{if TRUE, some gestures are combined (e.g., HittingSoft, HitSoft, and Hitting become 'Hit)}
}
\value{
Function returns data frame with the observed probabilities for each Gesture/Goal combination, how often the gesture and goal occurred in total, how specific the Goal is to the Gesture [Specificity.gesture], how specific the Gesture is to the Goal [Specificity.goal], and the expected probabilities for the goal within the gesture and the gesture within the goal with p-values
}
\description{
Function uses Fisher's exact test of Goal/Gesture combinations to determine those with higher-than-expected occurrence
}
