% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_from_Com_number.R
\name{info_from_Com_number}
\alias{info_from_Com_number}
\title{Helper functions to extract info from Clip and Communication Numbers}
\usage{
info_from_Com_number(
  com_number_column,
  output = c("coder", "group", "species")
)
}
\arguments{
\item{com_number_column}{vector with Communication Numbers}

\item{output}{can be 'coder', 'group', or 'species'}
}
\value{
returns Date, ClipNr, Coder, Group, and Species as vectors
}
\description{
Helper functions to extract info from Clip and Communication Numbers
}
