% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifier_entropy.R
\name{modifier_entropy}
\alias{modifier_entropy}
\title{Function that goes through the modifiers one by one and tests whether they have any impact on the entropy of the distribution of a target variable}
\usage{
modifier_entropy(
  data,
  modifiers,
  gesture_action,
  plot.action,
  cutoff = 1,
  target
)
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{modifiers}{a vector with the names of the modifiers to be tested}

\item{gesture_action}{a vector with the column that the gesture actions are stored in}

\item{plot.action}{a string of the gesture action of interest}

\item{cutoff}{minimum number of times a modifier level should occur to be counted}

\item{target}{vector of the same length as the gesture_action variable}
}
\value{
Data frame with the modifier, their level, the number of occurrences, the ratio observed/expected entropy, and a p-value (proportion of expected entropy values that are smaller than the observed value)
}
\description{
Calculates the observed Shannon entropy of the target within a modifier level; then randomises the modifier x times and calculates the entropy for each iteration, and compares the observed and expected entropy values
We make the assumption that a modifier that has no impact on the entropy of the target might not contain relevant information
}
