% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_detection_bayes.R
\name{morph_detection_bayes}
\alias{morph_detection_bayes}
\title{Takes modifiers for one gesture action, performs Bayesian Latent Class Analysis, and identifies clusters of modifiers}
\usage{
morph_detection_bayes(data, modifiers, gesture_action, plot.action, cutoff = 1)
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{modifiers}{a vector with the names of the modifiers to be tested}

\item{gesture_action}{a vector with the column that the gesture actions are stored in}

\item{plot.action}{a string of the gesture action of interest}

\item{cutoff}{minimum number of times a modifier level should occur to be counted}
}
\value{
Function returns cluster plot, information about the possible cluster solutions for individual elements and bigrams and their specificity to that cluster, BIC values for the cluster solutions, plot of the BIC, and the modifier data that were used

If there are fewer than 3 cases or only one modifier level for the gesture action, there is no information given
}
\description{
LCA is used to find the cluster solution with the highest BIC
}
