% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_goal_prediction.R
\name{morph_goal_prediction}
\alias{morph_goal_prediction}
\title{Give R a prediction column and a target column and it will build a Naive Bayes classifier to predict the latter}
\usage{
morph_goal_prediction(predictor, target, out = 10)
}
\arguments{
\item{predictor}{vector with the predictor (e.g., gesture actions)}

\item{target}{vector with the outcome variable}

\item{out}{how many folds should the prediction be repeated on}
}
\value{
accuracy of the naive bayes and random
}
\description{
Give R a prediction column and a target column and it will build a Naive Bayes classifier to predict the latter
}
