% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bipartite.R
\name{plot_bipartite}
\alias{plot_bipartite}
\title{Plots the overlap of multiple gesture actions and modifiers as bipartite network}
\usage{
plot_bipartite(
  prob.table,
  select.modifier,
  plot.title = NULL,
  cutoff = 1,
  threshold = 0,
  remove.full = FALSE
)
}
\arguments{
\item{prob.table}{data frame as produced by the probability_table() function}

\item{select.modifier}{a vector with the names of the modifiers to be tested; if 'All', it selects all modifiers established in probability.table(); if 'cluster', it plots the outcome of gesture.action.umap()}

\item{plot.title}{title of the resulting network plot}

\item{cutoff}{minimum number of occurrences for which modifier or Gesture should be included}

\item{threshold}{minimum conditional probability from which modifier or Gesture should be included}

\item{remove.full}{remove all cases where a gesture only exibits one modifier level, and only keep ambiguous ones}
}
\value{
network plot of bipartite network containing modifiers and gestures; or modifiers within cluster solutions

for networks of cluster solutions, plot only modifier levels that have specificity higher than randomly expected
}
\description{
The function takes the probability table and selected modifiers and plots how different elements
modifiers connect the gesture actions, based on the conditional probabilities that the
modifier occurs in the gesture and that the gesture is seen when the
modifier is present
}
