% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_table.R
\name{probability_table}
\alias{probability_table}
\title{Takes the dataset and modifiers and calculates conditional and unconditional probabilities of all Gesture/Modifier Level combinations}
\usage{
probability_table(data, modifiers)
}
\arguments{
\item{data}{ata frame as extracted from Filemaker}

\item{modifiers}{a vector with the names of the modifiers to be tested}
}
\value{
data frame containing all gesture actions and modifier levels with their occurrences and probabilities
}
\description{
Basis for a couple of other functions
}
