% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_levels.R
\name{reduce_levels}
\alias{reduce_levels}
\title{Takes the dataset and reduces some levels for a number of modifiers to facilitate further analyses. Main removes multiple NA, NotApplicable, NoValue etc}
\usage{
reduce_levels(
  data,
  col_names = "renamed",
  reduce.all = FALSE,
  reduce.nas = TRUE,
  reduce.goals = FALSE,
  reduce.gestures = FALSE,
  reduce.directed.to = FALSE,
  reduce.goal.met = FALSE,
  reduce.body.part.signaller = FALSE,
  reduce.body.part.contact = FALSE,
  reduce.object.used = FALSE,
  reduce.object.contacted = FALSE,
  reduce.laterality = FALSE,
  reduce.effectiveness = FALSE,
  reduce.distance = FALSE,
  reduce.visual.attention = FALSE,
  reduce.location = FALSE,
  reduce.visibility = FALSE
)
}
\arguments{
\item{data}{data frame as extracted from Filemaker}

\item{col_names}{either 'renamed' or 'original'. If they are still original at this point, they will be renamed here}

\item{reduce.all}{if TRUE, you don't have to set all the others as TRUE}

\item{reduce.nas}{if TRUE, for most columns, NAs and NVs etc will be combined in NA}

\item{reduce.goals}{if TRUE, some goals will be combined (e.g., GroomMe and GroomYou become 'Groom')}

\item{reduce.gestures}{if TRUE, some gestures are combined (e.g., HittingSoft, HitSoft, and Hitting become 'Hit)}

\item{reduce.directed.to}{if TRUE, some Directed_towards will be combined}

\item{reduce.goal.met}{if TRUE, some Goal_met will be combined}

\item{reduce.body.part.signaller}{if TRUE, some Body_part_signaller will be combined}

\item{reduce.body.part.contact}{if TRUE, some Body_parts_contacted will be combined}

\item{reduce.object.used}{if TRUE, some Object_used will be combined}

\item{reduce.object.contacted}{if TRUE, some Object_contacted will be combined}

\item{reduce.laterality}{if TRUE, some Laterality will be combined}

\item{reduce.effectiveness}{if TRUE, some Effectiveness will be combined}

\item{reduce.distance}{if TRUE, some Distances will be combined}

\item{reduce.visual.attention}{if TRUE, some Visual_attention will be combined}

\item{reduce.location}{if TRUE, some Locations will be combined}

\item{reduce.visibility}{if TRUE, some Visibility will be combined}
}
\value{
data frame containing all data in cleaned fashion
}
\description{
Reduction of the following columns: Signaller Arousal, Recipient Arousal, Directed Towards, Goal_met, Directionality, Body Parts Signaller, Body Parts Contact, Flexion, Orientation, Object Used, Object contact, Laterality, Emphasis, Context Waiting, Effectiveness, Persistence, Sequences, Recipient Distance, Gaze, Audible, Recipient Reaction, Visual Attention, Location, Recipient Signals, Latency, Signaller Communication, Visibility, Behaviour change
}
