% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_goals_gestures_new.R
\name{similarity_goals_gestures_new}
\alias{similarity_goals_gestures_new}
\title{Calculate and plot the similarity of Gestures or Contexts based on the respective other column}
\usage{
similarity_goals_gestures_new(
  gesture_action,
  goal,
  parameter = "Goal",
  trials = 20,
  n_epochs = 5000,
  n_neighbours = 4
)
}
\arguments{
\item{gesture_action}{vector with the gesture actions}

\item{goal}{vector with the meaning variable}

\item{parameter}{parameter to be plotted; either 'Goal' (plots the distribution of goals) or 'Gesture' (plots the distribution of gestures)}

\item{trials}{number of times the cluster detection algorithm should be performed a robust solution}

\item{n_epochs}{how many iterations for the UMAP dimension reduction?}

\item{n_neighbours}{sets the n_neighbors argument in UMAP - the smaller, the more fine-grained the clustering algorithm solution is}
}
\value{
plots of the K-Means cluster solution with UMAP dimension reduction, dendrogram, and correspondence analysis. Data for dendrogram, cluster solutions, and collocation analysis
}
\description{
Uses two approaches: UMAP to reduce dimensions combined with K-Means clustering; Dendrograms; and Correspondence Analysis using FactoMineR
}
